Move environment to utils and compile command to tools
