+++ docs/source/tests/tools/binaries.rst
@@ -11,7 +11,7 @@ The tests for the :py:mod:`reprospect.tools.binaries.sass` subpackage can be fou
    binaries/sass
 
 .. automodule:: tests.tools.binaries.test_cuobjdump
-.. automodule:: tests.tools.binaries.test_demangle
+.. automodule:: tests.tools.binaries.test_demangler
 .. automodule:: tests.tools.binaries.test_elf
 .. automodule:: tests.tools.binaries.test_symtab
 .. automodule:: tests.tools.binaries.test_nvdisasm
