+++ docs/source/tests/utils.rst
@@ -8,15 +8,15 @@ Tests for :py:mod:`reprospect.utils`.
 
 .. automodule:: tests.utils.test_cmake
 
-Compile command
----------------
-
-.. automodule:: tests.utils.test_compile_command
-
 Detect GPUs
 -----------
 
 .. automodule:: tests.utils.test_detect
+
+Environment
+-----------
+
+.. automodule:: tests.utils.test_environment
 
 :code:`subprocess`
 ------------------
