+++ reprospect/tools/compile_command.py
@@ -15,7 +15,7 @@ def get_arch_from_compile_command(cmd: str) -> set[NVIDIAArch]:
     """
     Get NVIDIA architecture from compile command.
 
-    >>> from reprospect.utils.compile_command import get_arch_from_compile_command
+    >>> from reprospect.tools.compile_command import get_arch_from_compile_command
     >>> get_arch_from_compile_command('nvcc -arch=sm_89 test.cpp')
     {NVIDIAArch(family=<NVIDIAFamily.ADA: 'ADA'>, compute_capability=ComputeCapability(major=8, minor=9))}
     """
