+++ tests/testing/CMakeLists.txt
@@ -1,3 +1,1 @@
-add_pytest(environment)
-
 add_subdirectory(binaries)
