+++ tests/tools/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_pytest(architecture)
 add_pytest(cacher)
+add_pytest(compile_command)
 add_pytest(device_properties)
 
 add_subdirectory(binaries)
