+++ tests/utils/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_pytest(cmake)
-add_pytest(compile_command)
 add_pytest(detect)
+add_pytest(environment)
 add_pytest(subprocess_helpers)
 add_pytest(types)
